Tests/LibWeb/Text/expected/iframe-click-after-parent-scroll.txt

click at (10, 10)
